Migrate user preference pages
