+++ packages/frontend/pages/profile/user-preferences/dashboard-settings.tsx
@@ -1,0 +1,7 @@
+import { DashboardSettingsContainer } from '$container/user-preferences/dashboard-settings.container';
+
+const DashboardSettingsUserPreferencePage = () => {
+  return <DashboardSettingsContainer />;
+};
+
+export default DashboardSettingsUserPreferencePage;
