+++ packages/frontend/pages/profile/user-preferences/default-account-settings.tsx
@@ -1,0 +1,7 @@
+import { DefaultAccountSettingsContainer } from '$container/user-preferences/default-account-settings.container';
+
+const DefaultAccountSettingsUserPreferencePage = () => {
+  return <DefaultAccountSettingsContainer />;
+};
+
+export default DefaultAccountSettingsUserPreferencePage;
