+++ packages/frontend/pages/profile/user-preferences/index.tsx
@@ -1,0 +1,7 @@
+import { UserPreferences } from '$pages/profile/user-preferences/user-preferences';
+
+const UserPreferenceListingPage = () => {
+  return <UserPreferences />;
+};
+
+export default UserPreferenceListingPage;
