+++ packages/frontend/pages/profile/user-preferences/market-update-settings.tsx
@@ -1,0 +1,7 @@
+import { MarketUpdateSettingsContainer } from '$container/user-preferences/market-update-settings.container';
+
+const MarketUpdateSettingsUserPreferencePage = () => {
+  return <MarketUpdateSettingsContainer />;
+};
+
+export default MarketUpdateSettingsUserPreferencePage;
