+++ packages/frontend/pages/profile/user-preferences/maximum-items-per-page.tsx
@@ -1,0 +1,7 @@
+import { MaximumItemsPerPageContainer } from '$container/user-preferences/maximum-items-per-page.container';
+
+const MaximumItemsPerPageUserPreferencePage = () => {
+  return <MaximumItemsPerPageContainer />;
+};
+
+export default MaximumItemsPerPageUserPreferencePage;
