+++ packages/frontend/pages/profile/user-preferences/statistics-settings.tsx
@@ -1,0 +1,7 @@
+import { StatisticsSettingsContainer } from '$container/user-preferences/statistics-settings.container';
+
+const StatisticsSettingsUserPreferencePage = () => {
+  return <StatisticsSettingsContainer />;
+};
+
+export default StatisticsSettingsUserPreferencePage;
