+++ packages/frontend/src/pages/profile/user-preferences/preferences/user-dashboard-settings.tsx
@@ -1,4 +1,3 @@
-import { useRouter } from 'next/router';
 import { useEffect } from 'react';
 import { useForm } from 'react-hook-form';
 
@@ -9,67 +8,59 @@ import { CheckboxGroup } from '$elements/checkbox/checkbox.group';
 import { Heading } from '$elements/heading/heading';
 import { Loader } from '$elements/loader/loader';
 import { LoaderFullScreen } from '$elements/loader/loader.fullscreen';
-import {
-  useUpdateUserStatisticsSettings,
-  useUserStatisticsSettings,
-} from '$hooks/profile/user-preference/useStatisticsSettings';
 import { UpdatePageInfo } from '$renderers/seo/updatePageInfo';
 import { capitalize } from '$utils/capitalize';
 
 const allAccountTypes = Object.values(AccountTypeEnum);
 
-export interface UserStatisticsSettingsFormFields {
+export interface UserDashboardSettingsFormFields {
   accountTypes: AccountTypeEnum[];
 }
 
-export const UserStatisticsSettings = (): JSX.Element | null => {
-  const methods = useForm<UserStatisticsSettingsFormFields>();
+interface UserDashboardSettingsProps {
+  data?: UserDashboardSettingsFormFields;
+  isLoading: boolean;
+  isUpdating: boolean;
+  onSave: (data: UserDashboardSettingsFormFields) => Promise<void>;
+}
 
-  const { push } = useRouter();
-  const { data, isLoading: isLoadingDefault } = useUserStatisticsSettings();
-  const [setStatisticsSettings, { isLoading: isUpdating }] =
-    useUpdateUserStatisticsSettings();
-
-  const handleSave = async (
-    newUserStatisticsData: UserStatisticsSettingsFormFields
-  ) => {
-    await setStatisticsSettings({
-      accountTypes: newUserStatisticsData.accountTypes,
-    });
-
-    push('/profile/user-preferences');
-  };
+export const UserDashboardSettings = ({
+  data,
+  isLoading,
+  isUpdating,
+  onSave,
+}: UserDashboardSettingsProps): JSX.Element | null => {
+  const methods = useForm<UserDashboardSettingsFormFields>();
 
   useEffect(() => {
     methods.reset(data);
   }, [data, methods]);
 
-  const isLoading = isLoadingDefault;
-
   return (
     <>
       {isUpdating && <LoaderFullScreen />}
       <UpdatePageInfo
-        title="Statistics settings"
+        title="Dashboard settings"
         backLink={'/profile/user-preferences'}
       />
-      {isLoading && <Loader />}
-      {!isLoading && (
+      {isLoading ? (
+        <Loader />
+      ) : (
         <Form
           methods={methods}
-          onSubmit={handleSave}
+          onSubmit={onSave}
           submitLabel="Save"
           formFooterBackLink="/profile/user-preferences"
         >
           <Heading className="mb-4">Account types for stats and graph</Heading>
-          <CheckboxGroup testId="statistics-account-checkboxes">
+          <CheckboxGroup testId="dashboard-account-checkboxes">
             {allAccountTypes.map((type) => (
               <Checkbox
                 key={type}
                 id={type}
-                label={capitalize(type)}
                 value={type}
-                name="accountTypes"
+                label={capitalize(type)}
+                name={'accountTypes'}
               />
             ))}
           </CheckboxGroup>
